admin-server/requirements.txt
django>=2.0.7
django-debug-toolbar>=1.9.1
djangorestframework>=3.8.1
python-dateutil>=2.7.3
django-cors-headers>=2.4.0
openpyxl>=2.5.5
django-rest-swagger==2.1.2
djangorestframework_simplejwt==4.0.0
sqlalchemy
psycopg2
pymysql
image
